authorization for post put and delete

+++ FeedsApi/Controllers/FeedsController.cs
@@ -1,4 +1,6 @@
 ﻿using FeedApi.Data.Entities;
+using FeedsApi.DTOs;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -36,15 +38,73 @@ namespace FeedsApi.Controllers
             return feed;
         }
 
-        [HttpPut("{id}")]
-        public async Task<IActionResult> PutFeed(int id, Feed feed)
+        [Authorize]
+        [HttpPost]
+        public async Task<ActionResult<Feed>> PostFeed([FromBody] FeedDto feedDto)
         {
-            if (id != feed.FeedId)
+            var userId = GetLoggedInUserId();
+
+            Feed feed = feedDto.Discriminator switch
             {
-                return BadRequest();
+                "Image" => new ImageFeed
+                {
+                    Title = feedDto.Title,
+                    Description = feedDto.Description,
+                    ImageData = (feedDto as ImageFeedDto)?.ImageData
+                },
+                "Video" => new VideoFeed
+                {
+                    Title = feedDto.Title,
+                    Description = feedDto.Description,
+                    Url = (feedDto as VideoFeedDto)?.Url
+                },
+                _ => new Feed
+                {
+                    Title = feedDto.Title,
+                    Description = feedDto.Description
+                }
+            };
+
+            feed.UserId = userId;
+            _context.Feeds.Add(feed);
+            await _context.SaveChangesAsync();
+            return CreatedAtAction("GetFeed", new { id = feed.FeedId }, feed);
+        }
+
+        [Authorize]
+        [HttpPut("{id}")]
+        public async Task<IActionResult> PutFeed(int id, [FromBody] FeedDto feed)
+        {
+            var actualFeed = _context.Feeds.FirstOrDefault(f => f.FeedId == id);
+            //if not found, there's no feed with id
+            if (actualFeed == null)
+            {
+                return BadRequest("No feed with given id");
             }
 
-            _context.Entry(feed).State = EntityState.Modified;
+            //only update to same feed type
+            if (actualFeed.Discriminator != feed.Discriminator)
+            {
+                return BadRequest("Can't update to different type");
+            }
+
+            //check if the user is the owner of the feed
+            if (actualFeed.UserId != GetLoggedInUserId())
+            {
+                return Forbid();
+            }
+
+            actualFeed.Title = feed.Title;
+            actualFeed.Description = feed.Description;
+            if (actualFeed is ImageFeed imageFeed && feed is ImageFeedDto imageFeedDto)
+            {
+                imageFeed.ImageData = imageFeedDto.ImageData;
+            }
+            else if (actualFeed is VideoFeed videoFeed && feed is VideoFeedDto videoFeedDto)
+            {
+                videoFeed.Url = videoFeedDto.Url;
+            }
+            _context.Entry(actualFeed).State = EntityState.Modified;
 
             try
             {
@@ -65,14 +125,7 @@ namespace FeedsApi.Controllers
             return NoContent();
         }
 
-        [HttpPost]
-        public async Task<ActionResult<Feed>> PostFeed([FromBody] Feed feed)
-        {
-            _context.Feeds.Add(feed);
-            await _context.SaveChangesAsync();
-            return CreatedAtAction("GetFeed", new { id = feed.FeedId }, feed);
-        }
-
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteFeed(int id)
         {
@@ -80,6 +133,10 @@ namespace FeedsApi.Controllers
             if (feed == null)
             {
                 return NotFound();
+            }
+
+            if (feed.UserId != GetLoggedInUserId()) {
+                return Forbid();
             }
 
             _context.Feeds.Remove(feed);
@@ -91,6 +148,11 @@ namespace FeedsApi.Controllers
         private bool FeedExists(int id)
         {
             return _context.Feeds.Any(e => e.FeedId == id);
+        }
+
+        private int GetLoggedInUserId()
+        {
+            return int.Parse(HttpContext.User.Claims.First(claim => claim.Type.Contains("nameidentifier")).Value);
         }
     }
 }
